feat: Static site builder.
  - Add `vapid build` cli command.
  - Break out core /lib/Vapid.js file for config ingestion.
  - `lib/VapidServer.js` extends core `Vapid` class for dev server.
  - `lib/VapidBuilder.js` extends core `Vapid` class for static build.
  - Static builder generates static site under `/dist`.
  - Lots of Vapid specific path logic abstracted out to `lib/paths.js` util file.
  - Added `/dist` to template .gitignore.
  - Pulled template rendering functionality to `lib/renderer.js` util for sharing.
  - Abstracted some middleware business logic to path, renderer and favicon utils.
