+++ lib/dashboard.js
@@ -10,16 +10,14 @@ const LocalStrategy = require('passport-local');
 const Router = require('koa-router');
 const Sequelize = require('sequelize');
 
-const middleware = require('./middleware');
-const services = require('./services');
 const Form = require('./form');
+const middleware = require('./middleware');
+const { getDashboardPaths } = require('./paths');
+const services = require('./services');
 const Utils = require('./utils');
 
 const router = new Router({ prefix: '/dashboard' });
-const paths = {
-  assets: resolve(__dirname, '../assets'),
-  views: resolve(__dirname, '../views'),
-};
+const paths = getDashboardPaths();
 
 let local;
 let Record;
