+++ lib/middleware/assets.js
@@ -1,9 +1,7 @@
+const Boom = require('boom');
 const mount = require('koa-mount');
 const serve = require('koa-static');
-const { extname } = require('path');
-const Boom = require('boom');
-const Utils = require('../utils');
-
+const { isAssetPath } = require('../paths');
 /**
  * Looks for static assets
  *
@@ -15,15 +13,12 @@ const Utils = require('../utils');
  */
 module.exports = function assets(path, prefix = '/') {
   return async (ctx, next) => {
-    const ext = extname(ctx.path);
+    const stat = isAssetPath(ctx.path);
 
-    if (Utils.includes(['', '.html'], ext)) {
-      await next();
-    } else if (Utils.includes(['.scss', '.sass'], ext)) {
-      const suggestion = ctx.path.replace(/\.(scss|sass)$/, '.css');
-      throw Boom.notFound(`Sass files cannot be served. Use "${suggestion}" instead.`);
-    } else {
-      await mount(prefix, serve(path))(ctx, next);
-    }
+    // If it returns an error message, throw the error.
+    if (typeof stat === 'string') { throw Boom.notFound(stat); }
+
+    // Otherwise, serve or skip accordingly.
+    await (stat ? mount(prefix, serve(path))(ctx, next) : next());
   };
 };
