+++ lib/middleware/privateFiles.js
@@ -1,7 +1,4 @@
-const { basename } = require('path');
-const Boom = require('boom');
-
-const Utils = require('../utils');
+const { assertPublicPath } = require('../paths');
 
 /**
  * Throw 404 if the path starts with an underscore or period
@@ -13,12 +10,6 @@ const Utils = require('../utils');
  * @throws {Boom.notFound}
  */
 module.exports = async function privateFiles(ctx, next) {
-  const fileName = basename(ctx.path);
-  const char = fileName.slice(0, 1);
-
-  if (Utils.includes(['_', '.'], char)) {
-    throw Boom.notFound('Filenames starting with an underscore or period are private, and cannot be served.');
-  } else {
-    await next();
-  }
+  assertPublicPath(ctx.path);
+  await next();
 };
