+++ site_template/.gitignore.ejs
@@ -3,3 +3,4 @@
 data
 node_modules
 package-lock.json
+dist
